Use a str with urlsplit to avoid encode/decode confusion. Move common test code to the util.py and add a new test without $base to cover the changes here.
