+++ schema_salad/tests/test_java_codegen.py
@@ -2,10 +2,10 @@ import shutil
 from pathlib import Path
 from typing import Any, Dict, List, Optional, cast
 
-from schema_salad import codegen, ref_resolver
+from schema_salad import codegen
 from schema_salad.schema import load_schema
 
-from .util import get_data
+from .util import cwl_file_uri, metaschema_file_uri, get_data
 
 
 def test_cwl_gen(tmp_path: Path) -> None:
@@ -39,16 +39,6 @@ def test_meta_schema_gen(tmp_path: Path) -> None:
     assert pom_xml_path.exists()
     src_dir = target_dir / "src" / "main" / "java" / "org" / "w3id" / "cwl" / "salad"
     assert src_dir.exists()
-
-
-def get_data_uri(resource_path: str) -> str:
-    path = get_data(resource_path)
-    assert path
-    return ref_resolver.file_uri(path)
-
-
-cwl_file_uri = get_data_uri("tests/test_schema/CommonWorkflowLanguage.yml")
-metaschema_file_uri = get_data_uri("metaschema/metaschema.yml")
 
 
 def java_codegen(file_uri: str, target: Path, examples: Optional[Path] = None) -> None:
